Revert "new component folder added"
