+++ src/Main.js
@@ -7,7 +7,7 @@ import BookingHistory from "./pages/BookingHistory/BookingHistory.js";
 import VehicleList from "./pages/VehicleList/index.js";
 import VehicleDetails from "./pages/VechileDetails/index.js";
 import BookingConfirm from "./pages/BookingConfirm/index.js";
-import PreviewPage from "./pages/Preview/index.js";
+import Preview from "./pages/preview/index.js";
 import VehicleManagement from "./pages/VehicleManagement/index.js";
 import DeliveryManagement from "./pages/DeliveryManagement/index.js";
 import BookingDetails from "./pages/BookingDetails/index.js";
@@ -111,7 +111,11 @@ const Main = () => {
         />
         <Route
           path="/preview"
-          element={ <PreviewPage />
+          element={
+            <ProtectedRoute
+              element={<Preview />}
+              roles={["user"] || ["employee"]}
+            />
           }
         />
         <Route
