+++ src/components/Accordion/Accordion.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import "./Accordion.css";
 
-const Accordion = ({ header, defaultStatus = 'close', children }) => {
+const Accordion = ({ header, status, details, actions, children }) => {
   const [isExpanded, setIsExpanded] = useState(false);
 
   const toggleAccordion = () => {
@@ -12,6 +12,7 @@ const Accordion = ({ header, defaultStatus = 'close', children }) => {
     <div className="accordion">
       <div className="accordion-header">
         <div className="accordion-header-title">{header}</div>
+        <div>{status}</div>
         <button className="accordion-toggle-button" onClick={toggleAccordion}>
           {isExpanded ? "-" : "+"}
         </button>
@@ -19,7 +20,9 @@ const Accordion = ({ header, defaultStatus = 'close', children }) => {
       {isExpanded && (
         <div className="accordion-body">
           <div className="accordion-body-content">
+            <div className="accordion-details">{details}</div>
             {children}
+            <div className="accordion-actions">{actions}</div>
           </div>
         </div>
       )}
