--- src/newComponents/Accordion/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function NewAccordion(props) {
-    return (
-        <div>
-            New accordion component
-        </div>
-    );
-}
-
-export default NewAccordion;
