--- src/newComponents/Button/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function NewButton(props) {
-    return (
-        <div>
-            
-        </div>
-    );
-}
-
-export default NewButton;
